# docs/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ditto Quant API",
    "description": "量化投资系统API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/fundamental/financials/{report_type}": {
      "get": {
        "tags": [
          "fundamental"
        ],
        "summary": "Get Financials",
        "description": "获取财务报表数据.\n\n标识符三选一（优先级: instrument_id > standard_ticker > ticker）:\n- instrument_id: 内部 ID，如 1000001\n- standard_ticker: Ditto 标准格式，如 \"000001.XSHE\"\n- ticker: 裸代码，如 \"000001\"",
        "operationId": "get_financials_api_v1_fundamental_financials__report_type__get",
        "parameters": [
          {
            "name": "report_type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/FinancialType"
            }
          },
          {
            "name": "instrument_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Canonical 标的 ID",
              "title": "Instrument Id"
            },
            "description": "Canonical 标的 ID"
          },
          {
            "name": "ticker",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "裸代码, 如 000001",
              "title": "Ticker"
            },
            "description": "裸代码, 如 000001"
          },
          {
            "name": "standard_ticker",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "标准代码, 如 000001.XSHE",
              "title": "Standard Ticker"
            },
            "description": "标准代码, 如 000001.XSHE"
          },
          {
            "name": "as_of_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "PIT 查询日期",
              "title": "As Of Date"
            },
            "description": "PIT 查询日期"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_Financial__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fundamental/dividend": {
      "get": {
        "tags": [
          "fundamental"
        ],
        "summary": "Get Dividend",
        "description": "获取分红数据.\n\n标识符三选一（优先级: instrument_id > standard_ticker > ticker）:\n- instrument_id: 内部 ID，如 1000001\n- standard_ticker: Ditto 标准格式，如 \"000001.XSHE\"\n- ticker: 裸代码，如 \"000001\"",
        "operationId": "get_dividend_api_v1_fundamental_dividend_get",
        "parameters": [
          {
            "name": "instrument_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Canonical 标的 ID",
              "title": "Instrument Id"
            },
            "description": "Canonical 标的 ID"
          },
          {
            "name": "ticker",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "裸代码, 如 000001",
              "title": "Ticker"
            },
            "description": "裸代码, 如 000001"
          },
          {
            "name": "standard_ticker",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "标准代码, 如 000001.XSHE",
              "title": "Standard Ticker"
            },
            "description": "标准代码, 如 000001.XSHE"
          },
          {
            "name": "as_of_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "PIT 查询日期",
              "title": "As Of Date"
            },
            "description": "PIT 查询日期"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_Dividend__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fundamental/corporate-actions": {
      "get": {
        "tags": [
          "fundamental"
        ],
        "summary": "List Corporate Actions",
        "description": "查询公司行动列表.\n\n标识符三选一（优先级: instrument_id > standard_ticker > ticker）:\n- instrument_id: 内部 ID，如 1000001\n- standard_ticker: Ditto 标准格式，如 \"000001.XSHE\"\n- ticker: 裸代码，如 \"000001\"\n\nRaises:\n    HTTPException: 400 如果 start_date > end_date",
        "operationId": "list_corporate_actions_api_v1_fundamental_corporate_actions_get",
        "parameters": [
          {
            "name": "instrument_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Canonical 标的 ID",
              "title": "Instrument Id"
            },
            "description": "Canonical 标的 ID"
          },
          {
            "name": "ticker",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "裸代码, 如 000001",
              "title": "Ticker"
            },
            "description": "裸代码, 如 000001"
          },
          {
            "name": "standard_ticker",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "标准代码, 如 000001.XSHE",
              "title": "Standard Ticker"
            },
            "description": "标准代码, 如 000001.XSHE"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "开始日期",
              "title": "Start Date"
            },
            "description": "开始日期"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "结束日期",
              "title": "End Date"
            },
            "description": "结束日期"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_CorporateAction__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/metadata/instruments/{instrument_id}": {
      "get": {
        "tags": [
          "metadata"
        ],
        "summary": "Get Instrument",
        "description": "获取单个标的详情.\n\nArgs:\n    instrument_id: 标的 ID\n    facade: MetadataQueryFacade 依赖注入\n\nReturns:\n    APIResponse 包含 Instrument 标的信息\n\nRaises:\n    NotFoundError: 404 如果标的不存在",
        "operationId": "get_instrument_api_v1_metadata_instruments__instrument_id__get",
        "parameters": [
          {
            "name": "instrument_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instrument Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_Instrument_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/metadata/instruments": {
      "get": {
        "tags": [
          "metadata"
        ],
        "summary": "List Instruments",
        "description": "查询标的列表.\n\nArgs:\n    asset_class: 资产类别过滤 (可选)\n    exchange: 交易所过滤 (可选)\n    is_active: 活跃状态过滤 (可选)\n    pagination: 分页参数\n    facade: MetadataQueryFacade 依赖注入\n\nReturns:\n    APIResponse 包含 Instrument 列表 + 分页信息",
        "operationId": "list_instruments_api_v1_metadata_instruments_get",
        "parameters": [
          {
            "name": "asset_class",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AssetClass"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Class"
            }
          },
          {
            "name": "exchange",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exchange"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Limit"
            },
            "description": "每页数量"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "偏移量",
              "default": 0,
              "title": "Offset"
            },
            "description": "偏移量"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_Instrument__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponse_Instrument_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Instrument",
            "description": "响应数据"
          },
          "pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "分页信息(可选)"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIResponse[Instrument]"
      },
...
      "APIResponse_list_CorporateAction__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CorporateAction"
            },
            "type": "array",
            "title": "Data",
            "description": "响应数据"
          },
          "pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "分页信息(可选)"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIResponse[list[CorporateAction]]"
      },
      "APIResponse_list_Dividend__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Dividend"
            },
            "type": "array",
            "title": "Data",
            "description": "响应数据"
          },
          "pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "分页信息(可选)"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIResponse[list[Dividend]]"
      },
...
      "APIResponse_list_Financial__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Financial"
            },
            "type": "array",
            "title": "Data",
            "description": "响应数据"
          },
          "pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "分页信息(可选)"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIResponse[list[Financial]]"
      },
...
      "APIResponse_list_Instrument__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Instrument"
            },
            "type": "array",
            "title": "Data",
            "description": "响应数据"
          },
          "pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "分页信息(可选)"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIResponse[list[Instrument]]"
      },
...
      "AssetClass": {
        "type": "string",
        "enum": [
          "stock",
          "etf",
          "index",
          "future",
          "bond",
          "fund"
        ],
        "title": "AssetClass",
        "description": "资产类型枚举.\n\nAttributes:\n    STOCK: 股票\n    ETF: 交易所交易基金\n    INDEX: 指数\n    FUTURE: 期货\n    BOND: 债券\n    FUND: 基金"
      },
...
      "CorporateAction": {
        "properties": {
          "instrument_id": {
            "type": "integer",
            "title": "Instrument Id",
            "description": "标的 ID"
          },
          "action_date": {
            "type": "string",
            "title": "Action Date",
            "description": "行动日期"
          },
          "action_type": {
            "type": "string",
            "title": "Action Type",
            "description": "行动类型"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "行动描述"
          }
        },
        "type": "object",
        "required": [
          "instrument_id",
          "action_date",
          "action_type",
          "description"
        ],
        "title": "CorporateAction",
        "description": "公司行动响应模型.\n\nAttributes:\n    instrument_id: 标的 ID\n    action_date: 行动日期\n    action_type: 行动类型\n    description: 行动描述"
      },
...
      "Dividend": {
        "properties": {
          "instrument_id": {
            "type": "integer",
            "title": "Instrument Id",
            "description": "标的 ID"
          },
          "announce_date": {
            "type": "string",
            "title": "Announce Date",
            "description": "公告日期"
          },
          "dividend_type": {
            "type": "string",
            "title": "Dividend Type",
            "description": "分红类型"
          },
          "amount": {
            "type": "number",
            "title": "Amount",
            "description": "分红金额"
          }
        },
        "type": "object",
        "required": [
          "instrument_id",
          "announce_date",
          "dividend_type",
          "amount"
        ],
        "title": "Dividend",
        "description": "分红响应模型.\n\nAttributes:\n    instrument_id: 标的 ID\n    announce_date: 公告日期\n    dividend_type: 分红类型\n    amount: 分红金额"
      },
...
      "Financial": {
        "properties": {
          "instrument_id": {
            "type": "integer",
            "title": "Instrument Id",
            "description": "标的 ID"
          },
          "report_date": {
            "type": "string",
            "title": "Report Date",
            "description": "报告期"
          },
          "report_type": {
            "type": "string",
            "title": "Report Type",
            "description": "报表类型"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data",
            "description": "财务数据字典"
          }
        },
        "type": "object",
        "required": [
          "instrument_id",
          "report_date",
          "report_type",
          "data"
        ],
        "title": "Financial",
        "description": "财务报表响应模型.\n\nAttributes:\n    instrument_id: 标的 ID\n    report_date: 报告期\n    report_type: 报表类型\n    data: 财务数据字典"
      },
      "FinancialType": {
        "type": "string",
        "enum": [
          "balance_sheet",
          "income_statement",
          "cash_flow"
        ],
        "title": "FinancialType",
        "description": "财务报表类型枚举.\n\nAttributes:\n    BALANCE_SHEET: 资产负债表\n    INCOME_STATEMENT: 利润表\n    CASH_FLOW: 现金流量表"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Instrument": {
        "properties": {
          "instrument_id": {
            "type": "integer",
            "title": "Instrument Id",
            "description": "标的 ID"
          },
          "ticker": {
            "type": "string",
            "title": "Ticker",
            "description": "裸代码"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "证券名称"
          },
          "asset_class": {
            "$ref": "#/components/schemas/AssetClass",
            "description": "资产类别"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange",
            "description": "交易所代码"
          },
          "list_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "List Date",
            "description": "上市日期"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "是否活跃"
          }
        },
        "type": "object",
        "required": [
          "instrument_id",
          "ticker",
          "name",
          "asset_class",
          "exchange",
          "is_active"
        ],
        "title": "Instrument",
        "description": "标的响应模型.\n\nAttributes:\n    instrument_id: 标的 ID\n    ticker: 裸代码 (如 600000)\n    name: 证券名称\n    asset_class: 资产类别\n    exchange: 交易所代码\n    list_date: 上市日期 (可选)\n    is_active: 是否活跃"
      },
...
      "PaginationResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "description": "总记录数"
          },
          "limit": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Limit",
            "description": "每页数量"
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "description": "当前偏移量"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "判断是否还有更多数据.",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "PaginationResponse",
        "description": "分页响应模型.\n\nAttributes:\n    total: 总记录数.\n    limit: 每页数量.\n    offset: 当前偏移量.\n    has_more: 是否有更多数据(计算字段)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}